feat(sever): test template
