--- packages/create-twenty-app/src/constants/base-application/vitest.config.ts
@@ -1,20 +1,0 @@
-import tsconfigPaths from 'vite-tsconfig-paths';
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  plugins: [
-    tsconfigPaths({
-      root: __dirname,
-      ignoreConfigErrors: true,
-    }),
-  ],
-  test: {
-    testTimeout: 120_000,
-    hookTimeout: 120_000,
-    include: ['src/**/*.integration-test.ts'],
-    env: {
-      TWENTY_API_URL: 'http://localhost:3000',
-      TWENTY_CONFIG_PATH: '.twenty-test.config.json',
-    },
-  },
-});
