+++ packages/create-twenty-app/tsconfig.json
@@ -23,5 +23,8 @@
     "**/__mocks__/**/*",
     "vite.config.ts",
     "jest.config.mjs"
+  ],
+  "exclude": [
+    "src/constants/base-application/vitest.config.ts"
   ]
 }
